Emily's first changes
This is really great, really thorough content. For the code pattern on IBM Developer, I will probably just take everything from the Runtime flow heading down and use that as the code pattern. 

I think the content above that is really great, but it seems a little more like a tutorial. I wonder if we shoudl make it a separate piece of content and just say something like: New to test-driven development? Read these blogs for a thorough introduction. 

One thing, for the IBM Developer page, I will need to get a little bit of background about exactly what this code pattern is showing them how to do (besides TDD--what does the actual code do?).